case文に追加する方法をやめた
